+++ app/src/main/java/com/example/marvi/taller_listview/areas_Rectangulo.java
@@ -26,9 +26,9 @@ public class areas_Rectangulo extends AppCompatActivity {
 
     public void  AreaRectangulo(View v){
 
-        int b = Integer.parseInt(base.getText().toString());
-        int h = Integer.parseInt(altura.getText().toString());
-        int calculo_area;
+        double b = Double.parseDouble(base.getText().toString());
+        double h = Double.parseDouble(altura.getText().toString());
+        double calculo_area;
         Calculos p = new Calculos();
         p.CalcularArea_Rectangulo(b,h);
         p.getArea_Rectangulo();
